+++ aitoolman/__init__.py
@@ -3,7 +3,7 @@ from .client import LLMClient, LLMLocalClient
 from .channel import *
 from .model import *
 from .provider import LLMProviderManager
-from .util import load_config, load_config_str, get_id
+from .util import load_config, load_config_str, get_id, get_host_id
 from .resmanager import ResourceManager
 from .workflow import *
 from . import postprocess